feat UGN-204 - cleanup, added dynamic title, context for config
