+++ src/hooks/useRsi.ts
@@ -1,0 +1,4 @@
+import { useContext } from "react"
+import { RsiContext } from "../components/Providers"
+
+export const useRsi = () => useContext(RsiContext)
